chore(ci): Move classes to io.github.bric3.fireplace as required by OSSRH
See https://central.sonatype.org/publish/requirements/coordinates/
