--- fireplace-app/src/main/java/com/github/bric3/fireplace/DnDTarget.java
@@ -1,11 +1,0 @@
-package com.github.bric3.fireplace;
-
-import javax.swing.*;
-
-public interface DnDTarget {
-    JComponent getComponent();
-
-    void activate();
-
-    void deactivate();
-}
