+++ fireplace-app/src/main/java/io/github/bric3/fireplace/DnDTarget.java
@@ -1,0 +1,11 @@
+package io.github.bric3.fireplace;
+
+import javax.swing.*;
+
+public interface DnDTarget {
+    JComponent getComponent();
+
+    void activate();
+
+    void deactivate();
+}
